add: rename module to lingo (more catchy)
